% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabayesrisk.R
\name{betabayesrisk}
\alias{betabayesrisk}
\title{Bayesian risk of shinkage rule under beta prior.}
\usage{
betabayesrisk(alpha, a, b, m, s)
}
\arguments{
\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{a}{Shape parameter of the beta prior.}

\item{b}{Shape parameter of the beta prior.}

\item{m}{Upper value of the beta prior support.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
The bayesian risk value of the shrinkage rule.
}
\description{
Provides the bayesian risk of the wavelet shrinkage rule under beta prior.
}
\examples{
betabayesrisk(0.9,2,3,10,1)
}
