% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betashrink.R
\name{betashrink}
\alias{betashrink}
\title{Wavelet shrinkage under beta prior.}
\usage{
betashrink(d, alpha, a, b, m, s)
}
\arguments{
\item{d}{The empirical wavelet coefficients vector.}

\item{alpha}{The weight of the point mass at zero function of the prior.}

\item{a}{The shape parameter of the beta prior.}

\item{b}{The shape parameter of the beta prior.}

\item{m}{The upper value of the beta prior support.}

\item{s}{The standard deviation of the normal random noise.}
}
\value{
The shrunk wavelet coefficients vector.
}
\description{
Performs bayesian shrinkage under beta prior on empirical wavelet coefficients.
}
\examples{
betashrink(c(0.5,1,2),0.9,2,3,10,1)
}
