% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSHbayesrisk.R
\name{GSHbayesrisk}
\alias{GSHbayesrisk}
\title{Bayesian risk of shinkage rule under GSH prior.}
\usage{
GSHbayesrisk(alpha, tau, t, s)
}
\arguments{
\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{tau}{Scale parameter of the GSH prior.}

\item{t}{Kurtosis parameter of the GSH prior.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
Bayesian risk value of the shrinkage rule.
}
\description{
Provides the bayesian risk of the wavelet shrinkage rule under GSH prior.
}
\examples{
GSHbayesrisk(0.9,1,2,1)
}
