% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSHclasrisk.R
\name{GSHclasrisk}
\alias{GSHclasrisk}
\title{Classical risk of shinkage rule under GSH prior.}
\usage{
GSHclasrisk(theta, alpha, tau, t, s)
}
\arguments{
\item{theta}{Wavelet coefficients vector.}

\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{tau}{Scale parameter of the GSH prior.}

\item{t}{Kurtosis parameter of the GSH prior.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
Vector of classical risks of the shrinkage rule.
}
\description{
Provides the classical risk of the wavelet shrinkage rule under GSH prior.
}
\examples{
GSHclasrisk(c(0,1,2),0.9,1,2,1)
}
