% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSHshrink.R
\name{GSHshrink}
\alias{GSHshrink}
\title{Wavelet shrinkage under GSH prior.}
\usage{
GSHshrink(d, alpha, tau, t, s)
}
\arguments{
\item{d}{Empirical wavelet coefficients vector.}

\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{tau}{Scale parameter of the GSH prior.}

\item{t}{Kurtosis parameter of the GSH prior.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
Shrunk wavelet coefficients vector.
}
\description{
Performs bayesian shrinkage under GSH prior on empirical wavelet coefficients.
}
\examples{
GSHshrink(c(0.5,1,2),0.9,1,2,1)

}
