% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSHvar.R
\name{GSHvar}
\alias{GSHvar}
\title{Variance of shinkage rule under GSH prior.}
\usage{
GSHvar(theta, alpha, tau, t, s)
}
\arguments{
\item{theta}{Wavelet coefficients vector.}

\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{tau}{Scale parameter of the GSH prior.}

\item{t}{Kurtosis parameter of the GSH prior.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
Vector of variances of the shrinkage rule.
}
\description{
Provides the variance of the wavelet shrinkage rule under GSH prior.
}
\examples{
GSHvar(c(0,1,2),0.9,1,2,1)
}
