% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbayesrisk.R
\name{logbayesrisk}
\alias{logbayesrisk}
\title{Bayesian risk of shinkage rule under logistic prior.}
\usage{
logbayesrisk(alpha, t, s)
}
\arguments{
\item{alpha}{Weight of the point mass at zero function of the prior.}

\item{t}{Scale parameter of the logistic prior.}

\item{s}{Standard deviation of the normal random noise.}
}
\value{
Bayesian risk value of the shrinkage rule.
}
\description{
Provides the bayesian risk of the wavelet shrinkage rule under logistic prior.
}
\examples{
logbayesrisk(0.9,1,1)
}
