% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangshrink.R
\name{triangshrink}
\alias{triangshrink}
\title{Wavelet shrinkage under triangular prior.}
\usage{
triangshrink(d, alpha, m, s)
}
\arguments{
\item{d}{The empirical wavelet coefficients vector.}

\item{alpha}{The weight of the point mass at zero function of the prior.}

\item{m}{The upper value of the beta prior support.}

\item{s}{The standard deviation of the normal random noise.}
}
\value{
The shrunk wavelet coefficients vector.
}
\description{
Performs bayesian shrinkage under triangular prior on empirical wavelet coefficients.
}
\examples{
triangshrink(c(0.5,1,2),0.9,10,1)
}
